Criado módulo categories

+++ src/modules/players/players.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreatePlayerDto } from './dtos/create-player.dto';
@@ -9,7 +9,6 @@ import { UpdatePlayerDto } from './dtos/update-player.dto';
 @Injectable()
 export class PlayersService {
   private readonly logger = new Logger(PlayersService.name)
-  private readonly players: Player[] = []
 
   constructor(@InjectModel('Player') private readonly playerModel: Model<Player>) {
 
@@ -20,21 +19,58 @@ export class PlayersService {
   }
 
   async findPlayerByEmail(email: string): Promise<Player> {
-    return await this.playerModel.findOne({ email }).exec()
+    const player = await this.playerModel.findOne({ email }).exec()
+
+    if (!player) {
+      throw new NotFoundException(`Não existe nenhum jogador com o e-mail ${email}`)
+    }
+
+    return player
+  }
+
+  async findPlayerById(id: string): Promise<Player> {
+    const player = await this.playerModel.findById(id).exec()
+
+    if (!player) {
+      throw new NotFoundException('Jogador não encontrado')
+    }
+
+    return player
   }
 
   async createPlayer(createPlayerDto: CreatePlayerDto): Promise<Player> {
     this.logger.log(`Criando jogador => ${JSON.stringify(createPlayerDto)}`)
+    
+    const exists = (await this.playerModel.count({ email: createPlayerDto.email }) as number) > 0
+
+    if (exists) {
+      throw new BadRequestException(`Já existe um jogador com o e-mail ${createPlayerDto.email}`)
+    }
+
     return await new this.playerModel(createPlayerDto).save()
   }
 
-  async updatePlayer(id: string, updatePlayerDto: UpdatePlayerDto): Promise<Player> {
+  async updatePlayer(id: string, updatePlayerDto: UpdatePlayerDto): Promise<void> {
     this.logger.log(`Atualizando jogador => ${JSON.stringify(updatePlayerDto)}`)
-    return await this.playerModel.findByIdAndUpdate(id, updatePlayerDto).exec()
+    
+    const exists = (await this.playerModel.count({ _id: { $ne: id }, email: updatePlayerDto.email }) as number) > 0
+
+    if (exists) {
+      throw new BadRequestException(`Já existe um jogador com o e-mail ${updatePlayerDto.email}`)
+    }
+    
+    await this.playerModel.findByIdAndUpdate(id, updatePlayerDto).exec()
   }
 
   async deletePlayer(id: string): Promise<void> {
     this.logger.log(`Removendo jogador => ${id}`)
+    
+    const exists = (await this.playerModel.count({ _id: id }) as number) > 0
+
+    if (!exists) {
+      throw new NotFoundException('Jogador não encontrado')
+    }
+    
     await this.playerModel.findByIdAndRemove(id).exec()
   }
 }
